src/django-app/requirements.txt
asgiref==3.5.2
attrs==22.2.0
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-nested-admin==4.0.2
django-phonenumber-field==7.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-access-policy==1.3.0
drf-spectacular==0.24.2
drf-standardized-errors==0.12.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.1
phonenumberslite==8.12.56
Pillow==9.4.0
platformdirs==2.5.2
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
